move files to sensible sub folders
