+++ src/utils/domain/pageMerge.ts
@@ -1,5 +1,5 @@
 import type { Page, PagesData } from "./schemas";
-import { TOMBSTONE_RETENTION_MS } from "../constants";
+import { TOMBSTONE_RETENTION_MS } from "../../constants";
 
 export interface PageTombstone {
   pageId: string;
