+++ src/utils/hooks/usePageHistory.ts
@@ -1,8 +1,8 @@
 import { useEffect, useRef, useState, type Dispatch, type SetStateAction } from "react";
-import { saveToStorage, loadFromStorage } from "./storage";
-import { UNDO_STACK_KEY, REDO_STACK_KEY } from "../constants";
-import { reconcileHistoryAfterMerge, type PageTombstone } from "./pageMerge";
-import type { Page, PagesData } from "./schemas";
+import { saveToStorage, loadFromStorage } from "../storage/storage";
+import { UNDO_STACK_KEY, REDO_STACK_KEY } from "../../constants";
+import { reconcileHistoryAfterMerge, type PageTombstone } from "../domain/pageMerge";
+import type { Page, PagesData } from "../domain/schemas";
 
 // How long an edit burst (e.g. a run of keystrokes) can pause before the next
 // change starts a new undo step, so backspacing through a whole value only
